# api-gateway/docs/openapi.yml
openapi: 3.0.3
info:
  title: Event Booking API Gateway
  description: |
    A comprehensive API for managing events, bookings, and user authentication.
    This API uses NATS messaging for microservice communication and provides
    endpoints for user registration, event management, and booking operations.
  version: 1.0.0
  contact:
    name: API Support
    email: support@eventbooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.eventbooking.com
    description: Production server
paths:
...
  /api/events:
    get:
      tags:
        - Events
      summary: List all events
      description: Retrieves a list of all active events
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          description: Failed to retrieve events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Events
      summary: Create a new event
      description: Creates a new event (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event by ID
      description: Retrieves detailed information about a specific event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The event ID
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Events
      summary: Update an event
      description: Updates an existing event (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The event ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Delete an event
      description: Deletes an existing event (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The event ID
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Failed to delete event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/search/{searchTerm}:
    get:
      tags:
        - Events
      summary: Search events
      description: Search events by name, description, or category
      parameters:
        - name: searchTerm
          in: path
          required: true
          schema:
            type: string
          description: Search term
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          description: Invalid search term
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/events/organizer/{organizerId}:
    get:
      tags:
        - Events
      summary: Get events by organizer
      description: Retrieves all events created by a specific organizer
      security:
        - BearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
          description: The organizer ID
      responses:
        '200':
          description: Organizer events fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          description: Invalid organizer ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/bookings/event/{eventId}:
    get:
      tags:
        - Bookings
      summary: Get bookings for an event (Admin only)
      description: Retrieves all bookings for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          description: The event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, cancelled, pending]
          description: Filter by booking status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Event bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '400':
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "An error occurred"
        data:
          type: array
          items:
            type: object
          description: Additional error details (validation errors, etc.)

    # Authentication Schemas
...
    CreateEventRequest:
      type: object
      required:
        - name
        - description
        - venue
        - dateTime
        - capacity
        - price
        - category
      properties:
        name:
          type: string
          minLength: 1
          example: "Tech Conference 2024"
        description:
          type: string
          example: "Annual technology conference featuring latest innovations"
        venue:
          type: string
          example: "Convention Center, Downtown"
        dateTime:
          type: string
          format: date-time
          example: "2024-06-15T09:00:00.000Z"
        capacity:
          type: integer
          minimum: 1
          example: 500
        price:
          type: number
          minimum: 0
          example: 99.99
        category:
          type: string
          example: "Technology"
    UpdateEventRequest:
      type: object
      properties:
        updates:
          type: object
          properties:
            name:
              type: string
              example: "Updated Event Name"
            description:
              type: string
              example: "Updated description"
            venue:
              type: string
              example: "New Venue Location"
            dateTime:
              type: string
              format: date-time
              example: "2024-06-15T09:00:00.000Z"
            capacity:
              type: integer
              minimum: 1
              example: 600
            availableTickets:
              type: integer
              minimum: 0
              example: 450
            price:
              type: number
              minimum: 0
              example: 89.99
            category:
              type: string
              example: "Technology"
            isActive:
              type: boolean
              example: true
    Event:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439012"
        name:
          type: string
          example: "Tech Conference 2024"
        description:
          type: string
          example: "Annual technology conference featuring latest innovations"
        venue:
          type: string
          example: "Convention Center, Downtown"
        dateTime:
          type: string
          format: date-time
          example: "2024-06-15T09:00:00.000Z"
        capacity:
          type: integer
          example: 500
        availableTickets:
          type: integer
          example: 450
        price:
          type: number
          example: 99.99
        category:
          type: string
          example: "Technology"
        organizerId:
          type: string
          example: "507f1f77bcf86cd799439011"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    EventResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Event'
        message:
          type: string
          example: "Event retrieved successfully"
    EventListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        message:
          type: string
          example: "Events retrieved successfully"

    # Booking Schemas
...
    Booking:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439013"
        eventId:
          type: string
          example: "507f1f77bcf86cd799439012"
        userId:
          type: string
          example: "507f1f77bcf86cd799439011"
        ticketQuantity:
          type: integer
          example: 2
        totalAmount:
          type: number
          example: 199.98
        status:
          type: string
          enum: [confirmed, cancelled, pending]
          example: "confirmed"
        bookingDate:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        paymentStatus:
          type: string
          enum: [pending, completed, failed, refunded]
          example: "completed"
        paymentMethod:
          type: string
          example: "credit_card"
        event:
          type: object
          properties:
            name:
              type: string
              example: "Tech Conference 2024"
            venue:
              type: string
              example: "Convention Center, Downtown"
            dateTime:
              type: string
              format: date-time
              example: "2024-06-15T09:00:00.000Z"
        user:
          type: object
          properties:
            name:
              type: string
              example: "John Doe"
            email:
              type: string
              format: email
              example: "john@example.com"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
...
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 100
            pages:
              type: integer
              example: 10
        message:
          type: string
          example: "Bookings retrieved successfully"
...